Use slider for color temp instead of wheel
