--- gallery/src/pages/components/ha-temp-color-picker.markdown
@@ -1,3 +1,0 @@
----
-title: Temp Color Picker
----
